chore: Stub out new test cards
